+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Added Archive APIsAdded Archive APIs
+
 ## 1.0.2
 
 - Added limit, orderBy, orderDescendingBy, offset in get and getByConditions APIs